+++ nautobot_circuit_maintenance/handle_notifications/handler.py
@@ -20,9 +20,11 @@ from nautobot_circuit_maintenance.models import (
 from .sources import get_notifications, MaintenanceNotification
 
 
+name = "Circuit Maintenance"  # pylint: disable=invalid-name
+
+
 # pylint: disable=broad-except
 PLUGIN_SETTINGS = settings.PLUGINS_CONFIG.get("nautobot_circuit_maintenance", {})
-MAX_INITIAL_DAYS_SINCE = 365
 
 
 def create_circuit_maintenance(
@@ -164,21 +166,20 @@ def create_or_update_circuit_maintenance(
     maintenance_id = f"{raw_entry.provider.slug}-{parser_maintenance.maintenance_id}"
     try:
         circuit_maintenance_entry = CircuitMaintenance.objects.get(name=maintenance_id)
-        # Using the RawNotification.date as the reference to sort because it's the one that takes into account the
-        # source receving time. The ParsedNotification.date stores the date when the RawNotification was parsed and the
-        # ParsedNotification was created.
+        # Using the RawNotification.stamp as the reference to sort because it's the one that takes into account the
+        # source receving time.
         last_parsed_notification = (
-            circuit_maintenance_entry.parsednotification_set.order_by("raw_notification__date").reverse().last()
+            circuit_maintenance_entry.parsednotification_set.order_by("raw_notification__stamp").reverse().last()
         )
 
         # If the notification is older than the latest one used to update the CircuitMaintenance, we skip updating it
         parser_maintenance_datetime = datetime.datetime.fromtimestamp(
             parser_maintenance.stamp, tz=datetime.timezone.utc
         )
-        if last_parsed_notification and last_parsed_notification.date > parser_maintenance_datetime:
+        if last_parsed_notification and last_parsed_notification.last_updated > parser_maintenance_datetime:
             logger.log_debug(
                 f"Not updating CircuitMaintenance {maintenance_id} because the notification is from "
-                f"{parser_maintenance_datetime}, older than the most recent notification from {last_parsed_notification.date}."
+                f"{parser_maintenance_datetime}, older than the most recent notification from {last_parsed_notification.last_updated}."
             )
             return circuit_maintenance_entry
 
@@ -225,7 +226,7 @@ def create_raw_notification(logger: Job, notification: MaintenanceNotification, 
         raw_entry = RawNotification.objects.get(
             subject=notification.subject,
             provider=provider,
-            date=parser.parse(notification.date),
+            stamp=parser.parse(notification.date),
         )
         # If the RawNotification was already created, we ignore it.
         if logger.debug:
@@ -239,7 +240,7 @@ def create_raw_notification(logger: Job, notification: MaintenanceNotification, 
                 raw=notification.raw_payload,
                 sender=notification.sender,
                 source=NotificationSource.objects.filter(name=notification.source).last(),
-                date=parser.parse(notification.date),
+                stamp=parser.parse(notification.date),
             )
             raw_entry.save()
             logger.log_success(raw_entry, message="Raw notification created.")
@@ -304,14 +305,14 @@ def process_raw_notification(logger: Job, notification: MaintenanceNotification)
 
 
 def get_since_reference(logger: Job) -> int:
-    """Get the timestamp from the latest processed RawNotification or a reference from config `initial_days_since`."""
+    """Get the timestamp from the latest processed RawNotification or a reference from config `raw_notification_initial_days_since`."""
     # Latest retrieved notification will limit the scope of notifications to retrieve
     last_raw_notification = RawNotification.objects.last()
     if last_raw_notification:
-        since_reference = last_raw_notification.date.timestamp()
+        since_reference = last_raw_notification.last_updated.timestamp()
     else:
-        since_reference = datetime.datetime.utcnow() - datetime.timedelta(
-            days=PLUGIN_SETTINGS.get("raw_notifications", {}).get("initial_days_since", MAX_INITIAL_DAYS_SINCE)
+        since_reference = datetime.datetime.now(datetime.timezone.utc) - datetime.timedelta(
+            days=PLUGIN_SETTINGS.get("raw_notification_initial_days_since")
         )
         since_reference = int(since_reference.timestamp())
     logger.log_info(message=f"Processing notifications since {since_reference}")