+++ nautobot_circuit_maintenance/models.py
@@ -1,10 +1,11 @@
 """Models for Circuit Maintenance."""
+import logging
 import pickle  # nosec
+from datetime import datetime, timezone
 from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.db import models
 from django.urls import reverse
-from django.utils.timezone import now
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 
@@ -14,8 +15,9 @@ from nautobot.core.models.generics import PrimaryModel, OrganizationalModel
 
 from .choices import CircuitImpactChoices, CircuitMaintenanceStatusChoices, NoteLevelChoices
 
+logger = logging.getLogger(__name__)
+
 PLUGIN_SETTINGS = settings.PLUGINS_CONFIG.get("nautobot_circuit_maintenance", {})
-DEFAULT_RAW_NOTIFICATION_SIZE = 1000
 
 
 @extras_features(
@@ -23,7 +25,7 @@ DEFAULT_RAW_NOTIFICATION_SIZE = 1000
     "custom_links",
     "custom_validators",
     "export_templates",
-    # "graphql",  # TODO: https://github.com/nautobot/nautobot/issues/399
+    "graphql",
     "relationships",
     "statuses",
     "webhooks",
@@ -79,7 +81,7 @@ class CircuitMaintenance(PrimaryModel):
     "custom_links",
     "custom_validators",
     "export_templates",
-    # "graphql",  # TODO: https://github.com/nautobot/nautobot/issues/399
+    "graphql",
     "relationships",
     "webhooks",
 )
@@ -136,12 +138,11 @@ class Note(OrganizationalModel):
         blank=True,
     )
     comment = models.TextField()
-    date = models.DateTimeField(default=now)
 
     csv_headers = ["maintenance", "title", "level", "comment"]
 
     class Meta:  # noqa: D106 "Missing docstring in public nested class"
-        ordering = ["date"]
+        ordering = ["last_updated"]
         unique_together = ["maintenance", "title"]
 
     def __str__(self):
@@ -154,7 +155,7 @@ class Note(OrganizationalModel):
 
     def to_csv(self):
         """Return fields for bulk view."""
-        return (self.maintenance, self.title, self.level, self.comment, self.date)
+        return (self.maintenance, self.title, self.level, self.comment, self.last_updated)
 
 
 @extras_features(
@@ -240,19 +241,17 @@ class RawNotification(OrganizationalModel):
     sender = models.CharField(max_length=200, default="", null=True, blank=True)
     source = models.ForeignKey(NotificationSource, on_delete=models.SET_NULL, null=True)
     parsed = models.BooleanField(default=False, null=True, blank=True)
-    # RawNotification.date is the date when the RawNotification was received by the Source
-    date = models.DateTimeField(default=now)
+    # RawNotification.stamp is the date when the RawNotification was received by the Source
+    stamp = models.DateTimeField()
 
     class Meta:  # noqa: D106 "Missing docstring in public nested class"
-        ordering = ["date"]
-        unique_together = ("date", "provider", "subject")
+        ordering = ["stamp"]
+        unique_together = ("stamp", "provider", "subject")
 
     def save(self, *args, **kwargs):
         """Custom save for RawNotification."""
         # Limiting the size of the notification stored.
-        self.raw = self.raw[
-            : PLUGIN_SETTINGS.get("raw_notifications", {}).get("raw_notification_size", DEFAULT_RAW_NOTIFICATION_SIZE)
-        ]
+        self.raw = self.raw[: PLUGIN_SETTINGS.get("raw_notification_size")]
         super().save(*args, **kwargs)
 
     def __str__(self):
@@ -265,7 +264,14 @@ class RawNotification(OrganizationalModel):
 
     def to_csv(self):
         """Return fields for bulk view."""
-        return (self.subject, self.provider, self.sender, self.source, self.raw, self.date, self.parsed)
+        return (self.subject, self.provider, self.sender, self.source, self.raw, self.stamp, self.parsed)
+
+    def clean(self):
+        """Add validation when creating a RawNotification."""
+        super().clean()
+
+        if self.stamp > datetime.now(timezone.utc):
+            logger.warning("Stamp time %s is not consistent, it's in the future.", self.stamp)
 
 
 @extras_features(
@@ -282,11 +288,9 @@ class ParsedNotification(OrganizationalModel):
     maintenance = models.ForeignKey(CircuitMaintenance, on_delete=models.CASCADE, default=None)
     raw_notification = models.ForeignKey(RawNotification, on_delete=models.CASCADE, default=None)
     json = models.JSONField()
-    # ParsedNotification.date is the date when after parsing a RawNotification, the ParsedNotification was created
-    date = models.DateTimeField(default=now)
 
     class Meta:  # noqa: D106 "Missing docstring in public nested class"
-        ordering = ["date"]
+        ordering = ["last_updated"]
 
     def __str__(self):
         """String value for HTML rendering."""
@@ -298,4 +302,4 @@ class ParsedNotification(OrganizationalModel):
 
     def to_csv(self):
         """Return fields for bulk view."""
-        return (self.maintenance, self.raw_notification, self.json, self.date)
+        return (self.maintenance, self.raw_notification, self.json, self.last_updated)